refactor(foundation): move StateMachine family to foundation (A3)
Phase 6 / Sub-phase A3: Move the StateMachine cluster (rules, sub_scenario,
visual_layout) from intelligent_behavior to mockforge-foundation::state_machine.

These types are pure data with only `rand` for state selection — no other
crate dependencies. They're used by:
- mockforge-core (graph/builder.rs)
- mockforge-scenarios (manifest, state_machine, vbr_integration)
- mockforge-vbr (entities)
- Various deprecated modules

Moved to mockforge-foundation::state_machine::{rules, sub_scenario, visual_layout}:
- ConsistencyRule, EvaluationContext, RuleAction, StateMachine, StateTransition
- SubScenario
- VisualLayout, VisualNode, VisualEdge, Viewport

Core re-exports preserved via thin shims in:
- src/intelligent_behavior/rules.rs
- src/intelligent_behavior/sub_scenario.rs
- src/intelligent_behavior/visual_layout.rs

Consumers migrated to import from foundation:
- mockforge-vbr/src/entities.rs
- mockforge-scenarios/src/manifest.rs, state_machine.rs, vbr_integration.rs,
  tests/state_machine_tests.rs

27 foundation tests passing (up from 18). Workspace builds clean.
