+++ crates/mockforge-vbr/Cargo.toml
@@ -16,6 +16,7 @@ categories.workspace = true
 [dependencies]
 # Core dependencies
 mockforge-core = { version = "0.3.70", path = "../mockforge-core" }
+mockforge-foundation = { path = "../mockforge-foundation" }
 mockforge-data = { version = "0.3.70", path = "../mockforge-data" }
 mockforge-http = { version = "0.3.70", path = "../mockforge-http" }
